+++ lib/data/board/recruit/repositories/recruit_detail_repository_impl.dart
@@ -3,7 +3,7 @@ import 'package:dongsoop/data/board/recruit/data_sources/recruit_detail_data_sou
 import 'package:dongsoop/data/board/recruit/models/recruit_detail_model.dart';
 import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_detail_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 
 class RecruitDetailRepositoryImpl implements RecruitDetailRepository {
   final RecruitDetailDataSource dataSource;
