+++ lib/data/board/recruit/repositories/recruit_list_repository_impl.dart
@@ -3,7 +3,7 @@ import 'package:dongsoop/data/board/recruit/data_sources/recruit_list_data_sourc
 import 'package:dongsoop/data/board/recruit/models/recruit_list_model.dart';
 import 'package:dongsoop/domain/board/recruit/entities/recruit_list_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_list_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 
 class RecruitListRepositoryImpl implements RecruitListRepository {
   final RecruitListDataSource dataSource;
