+++ lib/data/board/recruit/repositories/recruit_write_repository_impl.dart
@@ -2,7 +2,7 @@ import 'package:dongsoop/core/exception/exception.dart';
 import 'package:dongsoop/data/board/recruit/data_sources/recruit_write_data_source.dart';
 import 'package:dongsoop/domain/board/recruit/entities/recruit_write_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_write_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 
 class RecruitWriteRepositoryImpl implements RecruitWriteRepository {
   final RecruitWriteDataSource dataSource;
