--- lib/domain/board/recruit/repositories/recruit_detail_repository.dart
@@ -1,9 +1,0 @@
-import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
-import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-
-abstract class RecruitDetailRepository {
-  Future<RecruitDetailEntity> fetchRecruitDetail({
-    required int id,
-    required RecruitType type,
-  });
-}
