--- lib/domain/board/recruit/repositories/recruit_write_repository.dart
@@ -1,9 +1,0 @@
-import 'package:dongsoop/domain/board/recruit/entities/recruit_write_entity.dart';
-import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-
-abstract class RecruitWriteRepository {
-  Future<void> submitRecruitPost({
-    required RecruitType type,
-    required RecruitWriteEntity entity,
-  });
-}
