+++ lib/domain/board/recruit/use_cases/recruit_detail_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:dongsoop/domain/board/recruit/entities/recruit_detail_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_detail_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 
 class RecruitDetailUseCase {
   final RecruitDetailRepository repository;
