+++ lib/domain/board/recruit/use_cases/recruit_list_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:dongsoop/domain/board/recruit/entities/recruit_list_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_list_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 
 class RecruitListUseCase {
   final RecruitListRepository repository;
