+++ lib/domain/board/recruit/use_cases/recruit_write_use_case.dart
@@ -1,6 +1,6 @@
 import 'package:dongsoop/domain/board/recruit/entities/recruit_write_entity.dart';
 import 'package:dongsoop/domain/board/recruit/enum/recruit_types.dart';
-import 'package:dongsoop/domain/board/recruit/repositories/recruit_write_repository.dart';
+import 'package:dongsoop/domain/board/recruit/repositories/recruit_repository.dart';
 import 'package:dongsoop/main.dart';
 
 class RecruitWriteUseCase {
